+++ examples/example.js
@@ -1,3 +1,4 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const QuickMongo = require('../dist/index')
 
 const db = new QuickMongo({